+++ mathics/builtin/arithmetic.py
@@ -137,7 +137,16 @@ class _MPMathFunction(SympyFunction):
                 elif mpmath.isnan(result):
                     result = Symbol("Indeterminate")
                 else:
-                    result = from_mpmath(result)
+                    # FIXME: replace try/except as a context manager
+                    # like "with evaluation.from_mpmath()...
+                    # which can be instrumented for
+                    # or mpmath tracing and benchmarking on demand.
+                    # Then use it on other places where mpmath appears.
+                    try:
+                        result = from_mpmath(result)
+                    except OverflowError:
+                        evaluation.message("General", "ovfl")
+                        result = Expression("Overflow")
         else:
             prec = min_prec(*args)
             d = dps(prec)