+++ CHANGES.rst
@@ -5,6 +5,8 @@ Enhancements
 ============
 
 * ``D`` can now act over ``Integrate`` and  ``NIntegrate`` (fix issue #130).
+* numeric overflows now do not affect the full evaluation, but just the element
+  which produce it.
 * ``SameQ`` (``===``) handles chaining, e.g. ``a == b == c`` or ``SameQ[a, b, c]``
 * ``Simplify`` now has a semantics closer to the WMA, and handles properly expressions of the form
   ``Simplify[0^a]`` (issue #167)